Consolidate .netcoreapp.cs files because projects is no longer cross-compiled
