+++ src/libraries/System.Linq/tests/System.Linq.Tests.csproj
@@ -42,7 +42,6 @@
     <Compile Include="RepeatTests.cs" />
     <Compile Include="ReverseTests.cs" />
     <Compile Include="SelectManyTests.cs" />
-    <Compile Include="SelectManyTests.netcoreapp.cs" />
     <Compile Include="SelectTests.cs" />
     <Compile Include="SequenceEqualTests.cs" />
     <Compile Include="ShortCircuitingTests.cs" />
@@ -67,7 +66,6 @@
     <Compile Include="UnionTests.cs" />
     <Compile Include="WhereTests.cs" />
     <Compile Include="ZipTests.cs" />
-    <Compile Include="ZipTests.netcoreapp.cs" />
     <Compile Include="$(CommonTestPath)System\Linq\SkipTakeData.cs">
       <Link>Common\System\Linq\SkipTakeData.cs</Link>
     </Compile>
