+++ src/libraries/System.ObjectModel/tests/System.ObjectModel.Tests.csproj
@@ -32,6 +32,7 @@
     <Compile Include="System\ComponentModel\PropertyChangingEventArgsTests.cs" />
     <Compile Include="System\ComponentModel\TypeConverterAttributeTests.cs" />
     <Compile Include="System\ComponentModel\TypeDescriptionProviderAttributeTests.cs" />
+    <Compile Include="System\Windows\Markup\ValueSerializerAttributeTests.cs" />
     <Compile Include="KeyedCollection\TestMethods.cs" />
     <Compile Include="KeyedCollection\ConcreteTestClasses.cs" />
     <Compile Include="KeyedCollection\Utils.cs" />
@@ -41,11 +42,6 @@
     <Compile Include="$(CommonTestPath)System\Diagnostics\DebuggerAttributes.cs">
       <Link>Common\System\Diagnostics\DebuggerAttributes.cs</Link>
     </Compile>
-  </ItemGroup>
-  <ItemGroup Condition="'$(TargetsNetCoreApp)' == 'true'">
-    <Compile Include="KeyedCollection\TestMethods.netcoreapp.cs" />
-    <Compile Include="System\Collections\ObjectModel\KeyedCollectionTests.netcoreapp.cs" />
-    <Compile Include="System\Windows\Markup\ValueSerializerAttributeTests.cs" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="KeyedCollection\Serialization.cs" />
